Refactored header
Expanded the custom page titles. Refactored header files accordingly
