+++ mystic_tuner_frontend/src/app/page.jsx
@@ -21,7 +21,12 @@ export default async function Home() {
           />
         </div>
         <div className="flex flex-auto items-center justify-center col-start-2">
-          <Link href="/view-import-deck" className="flex rounded-3xl bg-sky-blue hover:bg-blue text-dark-grey m-2 p-2 w-60 h-20 justify-center items-center 
+          <Link 
+            href={{
+              pathname: "/view-import-deck/new-deck",
+              query: {title: "New Deck"}, 
+            }} 
+            className="flex rounded-3xl bg-sky-blue hover:bg-blue text-dark-grey m-2 p-2 w-60 h-20 justify-center items-center 
           text-5xl font-body no-underline">
             New Deck
           </Link>
