+++ mystic_tuner_frontend/src/components/header-components/header.jsx
@@ -1,4 +1,3 @@
-//Warning: DO NOT set to "useclient"; that will cause ourr authentication to be client sided
 import {auth0} from '@/lib/auth0.js';
 import LoginDropdown from './loginDropdown';
 import HomeButton from './homeButton';
