+++ mystic_tuner_frontend/src/components/header-components/pageTitle.jsx
@@ -10,9 +10,6 @@ export default function PageTitle(){
     //Collections of paths and Titles
     const titles = {
         "/": "Deck Tuning Assistant",
-        "/decklist": "My Decks",
-        "/import-new-deck": "Import New Deck",
-        "/view-user-decks": "My Decks"
     };
     //If the pageTitle is not in our pre-defined list, use the search parameter
     const pageTitle = (titles[pathName] ? titles[pathName] : searchParams.get('title'));
